Refactor admin page and inventory management UI
