+++ views/employeeview.php
@@ -3,9 +3,11 @@ session_start();
 $access = (int)$_SESSION['access'];
 if($access < 2) {
 	header('Location: 403.php');
+} else {
+	echo '<h1>This is the employee page</h1>'
+		.'<p>Your access level is: '.$_SESSION['access'].'</p>'
+		.'<p><a href="index.php">Web Store</a></p>'
+		.'<p><button onclick="checkout()">Checkout Customer</button></p>'
+		.'<p><button onclick="managestock()">Manage Stock</button></p>';
 }
-echo '<h1>This is the employee page</h1>';
-echo '<p>Your access level is: '.$_SESSION['access'].'</p>';
-echo '<p><a href="adminview.php">Admin Page</a></p>';
-echo '<p><a href="customerview.php">Customer Page</a></p>';
 ?>
