+++ views/index.php
@@ -4,8 +4,6 @@
 require('../lib/mvc.php');
 require(controllers.'/handler.php');
 require('cats.php'); 
-//session_start();
-//var_dump($_SESSION);
 ?>
 <title>Cool Store</title>
 <link rel="stylesheet" type="text/css" href="../views/css/shop.css">
